pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epilongai"
version = "0.1.0"
description = "Deep learning pipeline for Oxford Nanopore methylation data — preterm birth prediction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "EpiLongAI Team"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "polars>=0.20",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "pysam>=0.22",
    "pyfaidx>=0.7",
    "typer[all]>=0.9",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "tqdm>=4.65",
    "shap>=0.42",
    "captum>=0.6",
    "pydantic>=2.5",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.6",
]
workflow = [
    "snakemake>=8.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
    "snakemake>=8.0",
]

[project.scripts]
epilongai = "epilongai.cli:app"

[tool.setuptools.packages.find]
include = ["epilongai*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
